+++ vtm/src/org/oscim/tiling/source/mapfile/MapFileTileSource.java
@@ -31,7 +31,7 @@ import java.io.IOException;
 import java.io.RandomAccessFile;
 
 public class MapFileTileSource extends TileSource {
-    static final Logger log = LoggerFactory.getLogger(MapFileTileSource.class);
+    private static final Logger log = LoggerFactory.getLogger(MapFileTileSource.class);
 
     /**
      * Amount of cache blocks that the index cache should store.
@@ -44,7 +44,7 @@ public class MapFileTileSource extends TileSource {
     IndexCache databaseIndexCache;
     boolean experimental;
     File mapFile;
-    RandomAccessFile mInputFile;
+    private RandomAccessFile mInputFile;
 
     /**
      * The preferred language when extracting labels from this tile source.
@@ -53,7 +53,11 @@ public class MapFileTileSource extends TileSource {
     private Callback callback;
 
     public MapFileTileSource() {
-        super(0, 17);
+        this(0, 17);
+    }
+
+    public MapFileTileSource(int zoomMin, int zoomMax) {
+        super(zoomMin, zoomMax);
     }
 
     /**